ref: move nvim config to one file
