--- .config/nvim/lua/amirhossein5/lazy/colorizer.lua
@@ -1,6 +1,0 @@
-return {
-    'NvChad/nvim-colorizer.lua',
-    event = { "BufReadPost", "BufNewFile", "InsertEnter" },
-
-    config = true
-}
