--- .config/nvim/lua/amirhossein5/lazy/comment.lua
@@ -1,12 +1,0 @@
-return {
-    'numToStr/Comment.nvim',
-    dependencies = {
-        'JoosepAlviste/nvim-ts-context-commentstring'
-    },
-    event = { "BufReadPost", "BufNewFile", "InsertEnter" },
-    config = function()
-        require('Comment').setup({
-            pre_hook = require('ts_context_commentstring.integrations.comment_nvim').create_pre_hook(),
-        })
-    end
-}
