--- .config/nvim/lua/amirhossein5/lazy/harpoon.lua
@@ -1,26 +1,0 @@
-return {
-    "ThePrimeagen/harpoon",
-    branch = "harpoon2",
-    dependencies = { 'nvim-lua/plenary.nvim' },
-    event = { "BufReadPost", "BufNewFile", "InsertEnter" },
-    keys = {
-        "<leader>a", "<leader>w",
-        "<c-h>", "<c-t>", "<c-n>", "<c-m>", "<c-,>", "<c-.>",
-    },
-
-    config = function()
-        local harpoon = require("harpoon")
-
-        harpoon:setup()
-
-        vim.keymap.set("n", "<leader>a", function() harpoon:list():add() end)
-        vim.keymap.set("n", "<leader>w", function() harpoon.ui:toggle_quick_menu(harpoon:list()) end)
-
-        vim.keymap.set("n", "<C-h>", function() harpoon:list():select(1) end)
-        vim.keymap.set("n", "<C-t>", function() harpoon:list():select(2) end)
-        vim.keymap.set("n", "<C-n>", function() harpoon:list():select(3) end)
-        vim.keymap.set("n", "<C-m>", function() harpoon:list():select(4) end)
-        vim.keymap.set("n", "<C-,>", function() harpoon:list():select(5) end)
-        vim.keymap.set("n", "<C-.>", function() harpoon:list():select(6) end)
-    end
-}
