--- .config/nvim/lua/amirhossein5/lazy/init.lua
@@ -1,11 +1,0 @@
-return {
-    {
-        'tpope/vim-eunuch',
-        event = { "BufReadPost", "BufNewFile", "InsertEnter" },
-    },
-
-    {
-        'tpope/vim-sleuth',
-        event = { "BufReadPost", "BufNewFile", "InsertEnter" },
-    },
-}
