--- .config/nvim/lua/amirhossein5/lazy/undotree.lua
@@ -1,8 +1,0 @@
-return {
-    'mbbill/undotree',
-    keys = { "<leader>u" },
-
-    config = function()
-        vim.keymap.set("n", "<leader>u", vim.cmd.UndotreeToggle)
-    end
-}
