--- .config/nvim/lua/amirhossein5/lazy/vim-test.lua
@@ -1,17 +1,0 @@
-return {
-    'vim-test/vim-test',
-    keys = { "<leader>tn", "<leader>tf", "<leader>ts", "<leader>tl", "<leader>tw" },
-
-    config = function()
-        vim.cmd([[
-            nnoremap <Leader>tn :w<CR> :tabonly<CR> :silent TestNearest<CR>
-            nnoremap <Leader>tf :w<CR> :tabonly<CR> :TestFile<CR>
-            nnoremap <Leader>ts :w<CR> :tabonly<CR> :TestSuite<CR>
-            nnoremap <Leader>tl :w<CR> :tabonly<CR> :TestLast<CR>
-            nnoremap <Leader>tw :w<CR> :TestVisit<CR>
-
-            let g:test#basic#start_normal = 1
-            let test#php#phpunit#executable = 'phpunit'
-        ]])
-    end
-}
