--- .config/nvim/lua/amirhossein5/remap.lua
@@ -1,37 +1,0 @@
-vim.g.mapleader = ' '
-
-vim.keymap.set("v", "J", ":m '>+1<CR>gv=gv")
-vim.keymap.set("v", "K", ":m '<-2<CR>gv=gv")
-
-vim.keymap.set({ "n" }, "<leader>cn", ":cnext<CR>")
-vim.keymap.set({ "n" }, "<leader>cp", ":cprev<CR>")
-vim.keymap.set({ "n" }, "<leader>co", ":copen<CR>")
-vim.keymap.set({ "n" }, "<leader>cc", ":cclose<CR>")
-vim.cmd("autocmd BufReadPost quickfix nnoremap <buffer> <CR> <CR><C-w>o")
-
-vim.keymap.set("n", "J", "mzJ`z")
-vim.keymap.set("n", "<C-d>", "<C-d>zz")
-vim.keymap.set("n", "<C-u>", "<C-u>zz")
-vim.keymap.set("n", "<C-f>", "<C-f>zz")
-vim.keymap.set("n", "<C-b>", "<C-b>zz")
-vim.keymap.set("n", "n", "nzzzv")
-vim.keymap.set("n", "N", "Nzzzv")
-
-vim.keymap.set("x", "<leader>p", [["_dp]])
-vim.keymap.set("x", "<leader>P", [["_dP]])
-vim.keymap.set({ "n", "v" }, "<leader>y", [["+y]])
-vim.keymap.set({ "n", "v" }, "<leader>d", [["_d]])
-
-vim.keymap.set("i", "<C-c>", "<Esc>")
-vim.keymap.set("n", "<C-c>", ":nohls<CR>")
-
-vim.keymap.set("n", "Q", "<nop>")
-
-vim.keymap.set("n", "<leader>f", vim.lsp.buf.format)
-
-vim.keymap.set('n', "<leader>lv", ":e resources/views/<CR>")
-
-vim.keymap.set("n", "<leader>s", [[:%s/\<<C-r><C-w>\>/<C-r><C-w>/gc<Left><Left><Left>]])
-vim.keymap.set("v", "<leader>s", [[y:%s/<C-r>"//gc<Left><Left><Left>]])
-
-vim.keymap.set("n", "-", ":Ex<CR>")
