--- .config/nvim/lua/amirhossein5/set.lua
@@ -1,32 +1,0 @@
-vim.opt.guicursor = ""
-
-vim.opt.termguicolors = true
-
-vim.opt.wrap = false
-
-vim.opt.nu = true
-vim.opt.rnu = true
-
-vim.opt.tabstop = 4
-vim.opt.softtabstop = 4
-vim.opt.shiftwidth = 4
-vim.opt.expandtab = true
-vim.opt.smartindent = true
-
-vim.opt.swapfile = false
-vim.opt.backup = false
-vim.opt.undodir = os.getenv("HOME") .. "/.vim/undodir"
-vim.opt.undofile = true
-
-vim.opt.incsearch = true
-vim.opt.hlsearch = false
-
-vim.opt.scrolloff = 8
-
-vim.opt.signcolumn = "yes"
-vim.opt.colorcolumn = "80"
-
-vim.opt.updatetime = 50
-
-vim.g.netrw_browse_split = 0
-vim.g.netrw_banner = 0
